pyproject.toml
[project]
name = "mlops-lab1-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.9.0",
    "click>=8.3.0",
    "fastapi>=0.119.1",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "pillow>=12.0.0",
    "pylint>=4.0.2",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.38.0",
]
